Try to fix API

+++ src/main/java/com/axonactive/basketball/entities/Team.java
@@ -19,7 +19,7 @@ public class Team {
     private String name;
     private String location;
     private LocalDate dateFound;
-    private Integer salaryCap;
+    private Double salaryCap;
     @Enumerated(value = EnumType.STRING)
     private Conference conference;
     @JoinColumn
